Improvement: UAT feedback (#10)
Description and Motivation
This PR addresses the following UAT feedback:

- Fixed code to deal with non-DAGs graphs (e.g. CPDAG, PAG, etc)
- Added tests to check edge types are preserved
- Improved logging when extending a graph
- Added cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.add_time_edge method to add a time edge
between two nodes. This method allows to specify the time lag for source and destination variables. This avoids
having to create the corresponding node name manually or using the utility function cai_causal_graph.utils.get_name_with_lag.
- Changed the cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.extend_gragh method to work with
a non-negative backward_steps and forward_steps instead of strictly positive.
- Added cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.to_numpy_by_lag method to convert the
cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph to a dictionary of adjacency matrices where the
keys are the time lags with the values being the adjacency matrices with respect to the variables.

---------

Co-authored-by: GitHub Actions Bot <>