Improvement: graph consolidation (#12)
This PR make from dict more general in order to reuse code in other libraries.

Added the from_dict method for nodes and edges
Changed the behaviour when identifier and (time lag, variable name) are both provided. If identifier doesn't match the correct name, then it raises, otherwise it doesn't fail. I don't see why it should raise if the name is correct.

---------

Co-authored-by: GitHub Actions Bot <>
Co-authored-by: andrew-causalens <andrew@causalens.com>