Improvement: Add support for python 3.12 (#20)
Description and Motivation
Wanted to expand support to python 3.12 so we can start supporting in some of our internal packages. Still have 3.11 has the default version for the workflows.

Additional Context
Also removed the caching in dep checks as was done on other internal packages. This was due to issues with GitHub using cache for a specific python version on another version.

How Has This Been Tested?
Current unit tests. Added 3.12 to PR check right now to confirm it works.