Improvements: More general to and from methods (#58)
Description and Motivation
Updated to/from methods to ensure the type matches the calling class. There is also a small bug fix where the node variable type was not passed to the skeleton node.

Additional Context
Part of a wider effort to improve the generalisation of the base CausalGraph class.

How Has This Been Tested?
Extended unit tests. Also added some tests for JSON serialisability of Skeleton.

---------

Co-authored-by: GitHub Actions Bot <>