Improvement: Bound networkx to less than 3.3 (#66)
Description and Motivation
networkx 3.3 (release notes) claims to fix a bug (PR) on simple paths but it actually returns paths of length one when source and destination is the same. This is suboptimal behavior particularly if there is no self loop edge.

Additional Context
We will add 3.3 support in future but first need to understand impact on our other packages. Plus our base image for 2024 Q2 is using networkx 3.2.1.

How Has This Been Tested?
Ensure we are using networkx 3.2.1 in unit tests for python 3.11. The actions on this PR do that.