Repo: Update docs-builder to ~0.2.1 (#67)
Description and Motivation
The doc building and deploying was failing. Originally thought this was due to docs-builder but it is due to upgrade to poetry. Need to change syntax for specifying the causalens source in the Makefile.

How Has This Been Tested?
Pushed a dev tag to see if docs build and publish correctly. See here.

---------

Co-authored-by: GitHub Actions Bot <>