Bug: Possible to add Edges forward and backwards (#85)
## Description and Motivation
Added CausalGraphErrors.ReverseEdgeExistsError Exception, which allows to distinguish acyclicity and double-reverse edges. Introduced a check in the CausalGraph._set_edge() method.

## Additional Context
ADS discovered a bug that allowed the addition of directed and undirected edges between b and a to a graph with an existing undirected edge between a and b. 

## Commits
* Check if there is outgoing edge from destination

* Distinguish ReverseEdgeExistsError

* Update changelog.md

* Update changelog.md

* Formatting

* PR comments

* Address changelogs comments

---------

Co-authored-by: Damien van de Berg <damienvandeberg@Damiens-MacBook-Pro.local>